+++ app/ExternalFiles/Extensions/Real_Import.php
@@ -234,7 +234,7 @@ class Real_Import extends Extension_Base {
 
 		// detect count of URLs depending on slash at the end of the given URL.
 		$url_count = 1;
-		if( ! empty( $settings['urls'] ) & str_ends_with( $settings['urls'], '/' ) ) {
+		if( ! empty( $settings['urls'] ) && str_ends_with( $settings['urls'], '/' ) ) {
 			$url_count = 2;
 		}